# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HQuizlet Platform API
  version: 1.7.0
  description: |
    HQuizlet Platform API contract. All public application routes are exposed
    through the gateway at localhost:8080.
    Phase 9 adds File Upload: presigned URL upload flow for avatars, flashcard
    images, and study set thumbnails.
servers:
  - url: http://localhost:8080
    description: Local gateway
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ParticipantAuth:
      type: http
      scheme: bearer
      description: 'Opaque participant token issued on join. Use "Authorization: Bearer <participantToken>".'
  schemas:
    Error:
      type: object
      required: [code, message, requestId, details]
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "score must be between 0 and total"
        requestId:
          type: string
          example: "20260901143000.000000000"
        details:
          type: object
          additionalProperties: true
          example: {}
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        name: { type: string }
        image: { type: string }
        role: { type: string }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
        token: { type: string }
        expiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }
...
    PresignUploadRequest:
      type: object
      required: [upload_type, filename, content_type, file_size]
      properties:
        upload_type:
          type: string
          enum: [avatar, flashcard_image, study_set_thumbnail]
        filename: { type: string, minLength: 1, maxLength: 255 }
        content_type: { type: string }
        file_size: { type: integer, format: int64, minimum: 1 }
    PresignUploadResponse:
      type: object
      required: [file_id, upload_url, upload_method, expires_at, headers_required]
      properties:
        file_id: { type: string, format: uuid }
        upload_url: { type: string, format: uri }
        upload_method: { type: string, const: PUT }
        expires_at: { type: string, format: date-time }
        headers_required:
          type: object
          properties:
            Content-Type: { type: string }
...
    FileMetadata:
      type: object
      required: [file_id, upload_type, filename, content_type, size_bytes, url, status, created_at]
      properties:
        file_id: { type: string, format: uuid }
        upload_type: { type: string }
        filename: { type: string }
        content_type: { type: string }
        size_bytes: { type: integer, format: int64 }
        url: { type: string, format: uri }
        status: { type: string, enum: [pending, active, deleted] }
        created_at: { type: string, format: date-time }
        confirmed_at: { type: [string, 'null'], format: date-time }
    QuotaInfo:
      type: object
      required: [active_files, max_files, used_bytes, max_bytes]
      properties:
        active_files: { type: integer, minimum: 0 }
        max_files: { type: integer, const: 100 }
        used_bytes: { type: integer, format: int64, minimum: 0 }
        max_bytes: { type: integer, format: int64, const: 524288000 }
    FileListResponse:
      type: object
      required: [items, total, quota]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FileMetadata' }
        total: { type: integer, minimum: 0 }
        quota: { $ref: '#/components/schemas/QuotaInfo' }

    # ---- End Phase 9 Schemas ----

    # ---- End Phase 7 Schemas ----

    # ---- End Phase 6 Schemas ----
paths:
...
  /v1/files/presign:
    post:
      tags: [files]
      summary: Request a presigned URL for file upload
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PresignUploadRequest' }
      responses:
        '201':
          description: Presigned URL created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PresignUploadResponse' }
        '400':
          description: Invalid request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' } 
        '401': { description: Authentication required }
        '429':
          description: Quota exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/files/{fileId}:
    parameters:
      - { in: path, name: fileId, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [files]
      summary: Get file metadata (owner only)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileMetadata' }
        '401': { description: Authentication required }
        '403': { description: Not owner }
        '404': { description: File not found }
    delete:
      tags: [files]
      summary: Soft delete a file (owner only)
      security: [{ BearerAuth: [] }]
      responses:
        '204': { description: Deleted }
        '401': { description: Authentication required }
        '403': { description: Not owner }
        '404': { description: File not found }
...
  /v1/files:
    get:
      tags: [files]
      summary: List user files with quota info
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: per_page, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileListResponse' }
        '401': { description: Authentication required }

  # ---- End Phase 9 Endpoints ----

  # ---- End Phase 7 Endpoints ----
...
  /v1/auth/register:
    post:
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                name: { type: string }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { description: Invalid JSON, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Email already registered, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/auth/login:
    post:
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Authenticated, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '401': { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/auth/me:
    get:
      tags: [auth]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Current auth state, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
  /v1/auth/logout:
    post:
      tags: [auth]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Logged out }
  /v1/auth/logout-all:
    post:
      tags: [auth]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: All sessions revoked }
  /v1/auth/refresh:
    post:
      tags: [auth]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Refreshed session, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '401': { description: Invalid session, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/auth/profile:
    get:
      tags: [auth]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Profile, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
    patch:
      tags: [auth]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                image: { type: string }
      responses:
        '200': { description: Updated profile, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
...